+++ src/main/java/com/nucleusTeq/backend/services/Impl/CategoryServiceImp.java
@@ -2,9 +2,13 @@ package com.nucleusTeq.backend.services.Impl;
 
 
 import com.nucleusTeq.backend.dto.CategoryDTO;
+import com.nucleusTeq.backend.entities.Books;
 import com.nucleusTeq.backend.entities.Category;
+import com.nucleusTeq.backend.exception.ResourceNotFoundException;
 import com.nucleusTeq.backend.mapper.CategoryMapper;
+import com.nucleusTeq.backend.repositories.BooksRepository;
 import com.nucleusTeq.backend.repositories.CategoryRepository;
+import com.nucleusTeq.backend.repositories.IssuanceRepository;
 import com.nucleusTeq.backend.services.ICategoryService;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
@@ -24,6 +28,12 @@ public class CategoryServiceImp implements ICategoryService {
 
     @Autowired
     private CategoryRepository categoryRepository;
+
+    @Autowired
+    private BooksRepository booksRepository;
+
+    @Autowired
+    private IssuanceRepository issuanceRepository;
 
     @Override
     public List<CategoryDTO> fetchCategories() {
@@ -56,8 +66,37 @@ public class CategoryServiceImp implements ICategoryService {
 
     @Override
     public String deleteCategory(Long id) {
+
+        Optional<Category> category = categoryRepository.findById(id);
+        if (category.isEmpty()) {
+            return "Category not found";
+        }
+
+        List<Books> booksInCategory = booksRepository.findByCategoryId(id);
+
+        if (booksInCategory.isEmpty()) {
+            categoryRepository.deleteById(id);
+            return "Category deleted successfully";
+        }
+
+
+
+        boolean hasIssuedBooks = booksInCategory.stream()
+                        .anyMatch(book ->issuanceRepository.existsByBookIdAndStatus(book.getId(),"Issued"));
+
+        if (hasIssuedBooks) {
+
+            return "Category cannot be deleted as some books under this category are currently issued.";
+        }
+
+        booksInCategory.forEach(book ->
+                issuanceRepository.deleteByBookIdAndStatus(book.getId(), "Returned")
+        );
+
+        booksRepository.deleteAll(booksInCategory);
         categoryRepository.deleteById(id);
-        return "Category deleted successfully with ID: " + id;
+
+        return "Category and all related books deleted successfully with ID: " + id;
     }
 
     @Override
@@ -86,7 +125,38 @@ public class CategoryServiceImp implements ICategoryService {
     }
 
     @Override
-    public Category getCategoryByName(String name) {
-        return categoryRepository.findByName(name);
+    public String updateCategory(Long id,CategoryDTO categoryDTO) {
+
+        Optional<Category> optionalCategory = categoryRepository.findById(id);
+
+        if(optionalCategory.isPresent()) {
+            Category existingCategory  =  optionalCategory.get();
+            existingCategory.setCategoryDesc(categoryDTO.getCategoryDesc());
+            existingCategory.setName(categoryDTO.getName());
+
+            categoryRepository.save(existingCategory);
+
+            return  "Category updated successfully with ID: " + id;
+        }else {
+            throw  new ResourceNotFoundException("Book not fount with ID : "+ id);
+        }
+
     }
+
+    @Override
+    public Category getCategoryByName(String name) {
+        Category category = categoryRepository.findByName(name);
+        if (category != null) {
+            return category;
+        } else {
+            throw new ResourceNotFoundException("Category not found with name: " + name);
+        }
+    }
+
+    @Override
+    public long getCategoryCount() {
+        return categoryRepository.count();
+    }
+
+
 }
